Revert "1) Bug Fix (Change type column "ID" in table Users. Old type "Short", new type "Int"). 2) Add new tables (Groups, Labels, Priorities, Tasks, States, UserMaps)."
This reverts commit ee4adabcc8eeaeb10e8821550ba36d32edea60d1.

+++ PlanningManagmentData/dbo/Tables/Users.sql
@@ -1,5 +1,5 @@
 ﻿CREATE TABLE [dbo].[Users] (
-    [ID]       INT      IDENTITY (1, 1) NOT NULL,
+    [ID]       SMALLINT      IDENTITY (1, 1) NOT NULL,
     [Name]     VARCHAR (30)  NOT NULL,
     [Surname]  VARCHAR (30)  NOT NULL,
     [Sex]      VARCHAR (1)   NOT NULL,
